finished Gaussian blur implementation on the GPU

+++ cuda_calls.cu
@@ -4,8 +4,8 @@
 #include <tira/cuda/cudaGaussianFilter2D.cuh>
 
 
-glm::mat2* cudaGaussianBlur(glm::mat2* source, size_t width, size_t height, float sigma,
-                            size_t& out_width, size_t& out_height, int deviceID = 0) {
+glm::mat2* cudaGaussianBlur(glm::mat2* source, unsigned int width, unsigned int height, float sigma,
+                            unsigned int& out_width, unsigned int& out_height, int deviceID = 0) {
 
     cudaSetDevice(deviceID);
     glm::mat2* dest = GaussianFilter2D_cpu<glm::mat2>(source, width, height, sigma, sigma, out_width, out_height);
